requirements.txt
anyio==4.7.0
bleach==6.2.0
bokeh==3.6.2
certifi==2024.12.14
charset-normalizer==3.4.0
colorcet==3.1.0
contourpy==1.3.1
filelock==3.13.1
fsspec==2024.2.0
holoviews==1.20.0
hvplot==0.11.2
idna==3.10
Jinja2==3.1.5
kagglehub==0.3.6
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.2.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
pandas==2.2.3
panel==1.5.5
param==2.2.0
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
torch==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.3.0
watchfiles==1.0.3
webencodings==0.5.1
xyzservices==2024.9.0
